Rename HttpTraceContext to W3CTraceContextPropagator.
