+++ integration-tests/tracecontext/src/main/java/io/opentelemetry/Application.java
@@ -8,7 +8,7 @@ package io.opentelemetry;
 import com.google.gson.Gson;
 import io.opentelemetry.api.OpenTelemetry;
 import io.opentelemetry.api.trace.Span;
-import io.opentelemetry.api.trace.propagation.HttpTraceContext;
+import io.opentelemetry.api.trace.propagation.W3CTraceContextPropagator;
 import io.opentelemetry.context.Context;
 import io.opentelemetry.context.propagation.ContextPropagators;
 import io.opentelemetry.context.propagation.TextMapPropagator.Getter;
@@ -33,7 +33,8 @@ public class Application {
   private static final OpenTelemetry openTelemetry;
 
   static {
-    OpenTelemetry.setGlobalPropagators(ContextPropagators.create(HttpTraceContext.getInstance()));
+    OpenTelemetry.setGlobalPropagators(
+        ContextPropagators.create(W3CTraceContextPropagator.getInstance()));
     openTelemetry = OpenTelemetry.get();
   }
